+++ src/Pages/Home/Home.js
@@ -1,7 +1,7 @@
 import "../../Styles/Global.scss";
 import React, { useEffect, useState, useContext } from "react";
 import { getAllServices } from "../../Api/Api";
-import SearchBar from "../../Components/SearchBar";
+import SearchBar from "../../ReactComponents/SearchBar";
 import ServiceCard from "./ServiceCard";
 import { ContractorsContext } from "../../context/ContractorsContext";
 import { useNavigate } from "react-router-dom";
