+++ src/ReactComponents/FilterBar/logo.png
