+++ src/ReactComponents/Footer.js
