+++ src/ReactComponents/SearchResultCard.jsx
