+++ src/ReactComponents/StarRating/StarRating.js
